// src/generated/services/Cr22e_facturasdecompratisService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Cr22e_facturasdecompratisBase, Cr22e_facturasdecompratis } from '../models/Cr22e_facturasdecompratisModel';
import type { Cr22e_facturasdecompratisUploadColumnName } from '../models/Cr22e_facturasdecompratisModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Cr22e_facturasdecompratisService {
  private static readonly dataSourceName = 'cr22e_facturasdecompratis';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Cr22e_facturasdecompratisBase, 'cr22e_facturasdecompratiid'>): Promise<IOperationResult<Cr22e_facturasdecompratis>> {
    const result = await Cr22e_facturasdecompratisService.client.createRecordAsync<Omit<Cr22e_facturasdecompratisBase, 'cr22e_facturasdecompratiid'>, Cr22e_facturasdecompratis>(
      Cr22e_facturasdecompratisService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Cr22e_facturasdecompratisBase, 'cr22e_facturasdecompratiid'>>): Promise<IOperationResult<Cr22e_facturasdecompratis>> {
    const result = await Cr22e_facturasdecompratisService.client.updateRecordAsync<Partial<Omit<Cr22e_facturasdecompratisBase, 'cr22e_facturasdecompratiid'>>, Cr22e_facturasdecompratis>(
      Cr22e_facturasdecompratisService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Cr22e_facturasdecompratisService.client.deleteRecordAsync(
      Cr22e_facturasdecompratisService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Cr22e_facturasdecompratis>> {
    const result = await Cr22e_facturasdecompratisService.client.retrieveRecordAsync<Cr22e_facturasdecompratis>(
      Cr22e_facturasdecompratisService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Cr22e_facturasdecompratis[]>> {
    const result = await Cr22e_facturasdecompratisService.client.retrieveMultipleRecordsAsync<Cr22e_facturasdecompratis>(
      Cr22e_facturasdecompratisService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Cr22e_facturasdecompratis> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Cr22e_facturasdecompratisService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Cr22e_facturasdecompratisService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }

  public static async upload(id: string, columnName: Cr22e_facturasdecompratisUploadColumnName, file: File, fileDisplayName?: string): Promise<IOperationResult<void>> {
    const arrayBuffer = await file.arrayBuffer();
    const data = new Uint8Array(arrayBuffer);
    const result = await Cr22e_facturasdecompratisService.client.uploadFileToRecord(
      Cr22e_facturasdecompratisService.dataSourceName,
      id,
      columnName,
      fileDisplayName || file.name,
      data,
    );
    return result;
  }
}