// src/generated/models/Cr22e_razonsocialsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_razonsocialsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_razonsocialsstatecode = keyof typeof Cr22e_razonsocialsstatecode;
export const Cr22e_razonsocialsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_razonsocialsstatuscode = keyof typeof Cr22e_razonsocialsstatuscode;

export interface Cr22e_razonsocialsBase {
  cr22e_name: string;
  cr22e_nombrecorto?: string;
  cr22e_observaciones?: string;
  cr22e_razonsocialid: string;
  cr22e_rfc?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_razonsocialsstatecode;
  statuscode?: Cr22e_razonsocialsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_razonsocials extends Cr22e_razonsocialsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}