// src/generated/models/Cr22e_sucursalesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_sucursalesesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_sucursalesesstatecode = keyof typeof Cr22e_sucursalesesstatecode;
export const Cr22e_sucursalesesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_sucursalesesstatuscode = keyof typeof Cr22e_sucursalesesstatuscode;

export interface Cr22e_sucursalesesBase {
  cr22e_activo?: boolean;
  cr22e_name: string;
  cr22e_nombresucursal?: string;
  cr22e_sucursalesid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_sucursalesesstatecode;
  statuscode?: Cr22e_sucursalesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_sucursaleses extends Cr22e_sucursalesesBase {
  cr22e_activoname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}