// src/generated/models/TransactioncurrenciesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Transactioncurrenciescurrencytype = {
  0: 'System',
  1: 'Custom'
} as const;
export type Transactioncurrenciescurrencytype = keyof typeof Transactioncurrenciescurrencytype;
export const Transactioncurrenciesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Transactioncurrenciesstatecode = keyof typeof Transactioncurrenciesstatecode;
export const Transactioncurrenciesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Transactioncurrenciesstatuscode = keyof typeof Transactioncurrenciesstatuscode;

export interface TransactioncurrenciesBase {
  currencyname: string;
  currencyprecision: number;
  currencysymbol: string;
  currencytype: Transactioncurrenciescurrencytype;
  entityimage?: string;
  exchangerate: number;
  importsequencenumber?: number;
  isocurrencycode: string;
  overriddencreatedon?: string;
  statecode: Transactioncurrenciesstatecode;
  statuscode?: Transactioncurrenciesstatuscode;
  transactioncurrencyid: string;
}

export interface Transactioncurrencies extends TransactioncurrenciesBase {
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  currencytypename?: string;
  entityimage_timestamp?: number;
  entityimage_url?: string;
  entityimageid?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  statecodename?: string;
  statuscodename?: string;
  uniquedscid?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}
export type TransactioncurrenciesUploadColumnName = 'entityimage';